==> notebooks/exploration/2-Uniprot_RefSeq_Match.ipynb <==
# Script permettant de chercher l'existance de séquences à mismatch Uniprot dans la base locale RefSeq  
Input:  
    fasta contenant les séquence Uniprot d'intérêt  (my_Query)  
    fasta contenant toute les protéines présente dans RefSeq (my_DB)  
Output:  fichier match.txt une ligne par match entre les deux fichiers

# Deux fois la même fonction mais l'une retourne un dictionnaire de type titre:séquence l'autre retourne séquence:titre
def fasta2List(pathFasta):
    f = open(pathFasta, "r")
    title = []
    seq = []
    seq_temp = []
    for line in f:
        if line[0] == ">":
            seq.append(''.join(seq_temp).replace("\n", ""))
            title.append(line.replace("\n", ""))
            seq_temp = []
        else:
            seq_temp.append(line)
    seq.append(''.join(seq_temp).replace("\n", ""))
    seq.pop(0)
    dictionary = dict(zip(seq, title))
    return dictionary

def importQuery(pathFasta):
    f = open(pathFasta, "r")
    title = []
    seq = []
    seq_temp = []
    for line in f:
        if line[0] == ">":
            seq.append(''.join(seq_temp).replace("\n", ""))
            title.append(line.replace("\n", ""))
            seq_temp = []
        else:
            seq_temp.append(line)
    seq.append(''.join(seq_temp).replace("\n", ""))
    seq.pop(0)
    dictionary = dict(zip(title, seq))
    return dictionary


my_DB = fasta2List("/commun/bics/DB-Corentin/refseq-perfect/all.fasta")
my_Query = importQuery("refseq-perfect/ID_error3.fasta")

for i, j in my_Query:
    try: 
        match = my_DB[j]
        print("Match found: "+ my_DB[j]+" "+i)
    except:
        pass
